Changed `ReadModel` and `ReadModelList`
Moved them to a data module
