+++ sharedkernel/domain/__init__.py
@@ -1,6 +1,7 @@
-from . import errors, events, models, repositories, services
+from . import data, errors, events, models, repositories, services
 
 __all__ = [
+    "data",
     "errors",
     "events",
     "models",
